Feat/auth-improvements (#690)
* feat: on query TpJobseekerCv sub-objects, include userId=currentUser.id condition

* tmp

* simplify tpjobseekercv module and implement creation authorization checks

* restrict access to various endpoints

* complete auth checks

* fix(con): mutations typing

* fix(con): browse mentors language filter dropdown

* fix(tp): compile errors

* fix(con): use latest ConMentorshipMatch to determine current status between mentor and mentee

* feat(con): reset log mentoring session form, when form's modal opened

* fix(con): "wohoo new mentorship accepted" notification: make sure to check LATEST mentorship in the "should I show the notification or not" logic

* fix(con): misc query invalidation

* fix(tp/cv builder): re-render pdf only when isBusy toggles between true and false

+++ apps/nestjs-api/src/tp-jobseeker-cv/tp-jobseeker-cv-education-records/tp-jobseeker-cv-education-records.service.ts
@@ -1,11 +1,18 @@
-import { Injectable, NotFoundException } from '@nestjs/common'
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common'
 import {
   TpJobseekerCvEducationRecordEntity,
   TpJobseekerCvEducationRecordEntityProps,
   TpJobseekerCvEducationRecordMapper,
 } from '@talent-connect/common-types'
 import { deleteUndefinedProperties } from '@talent-connect/shared-utils'
-import { SfApiTpJobseekerCvEducationRecordsService } from '../salesforce-api/sf-api-tp-jobseeker-cv-education-records.service'
+import { CurrentUser } from '../../auth/current-user.decorator'
+import { CurrentUserInfo } from '../../auth/current-user.interface'
+import { SfApiTpJobseekerCvEducationRecordsService } from '../../salesforce-api/sf-api-tp-jobseeker-cv-education-records.service'
+import { TpJobseekerCvReadService } from '../tp-jobseeker-cv.read.service'
 import { TpJobseekerCvEducationRecordCreateInput } from './dtos/tp-jobseeker-cv-education-record-create.entityinput'
 import { TpJobseekerCvEducationRecordDeleteInput } from './dtos/tp-jobseeker-cv-education-record-delete.entityinput'
 import { TpJobseekerCvEducationRecordPatchInput } from './dtos/tp-jobseeker-cv-education-record-patch.entityinput'
@@ -14,7 +21,8 @@ import { TpJobseekerCvEducationRecordPatchInput } from './dtos/tp-jobseeker-cv-e
 export class TpJobseekerCvEducationRecordsService {
   constructor(
     private readonly api: SfApiTpJobseekerCvEducationRecordsService,
-    private readonly mapper: TpJobseekerCvEducationRecordMapper
+    private readonly mapper: TpJobseekerCvEducationRecordMapper,
+    private readonly cvReadService: TpJobseekerCvReadService
   ) {}
 
   async findAll(filter: any = {}) {
@@ -41,7 +49,17 @@ export class TpJobseekerCvEducationRecordsService {
     }
   }
 
-  async createFromInput(input: TpJobseekerCvEducationRecordCreateInput) {
+  async createFromInput(
+    input: TpJobseekerCvEducationRecordCreateInput,
+    @CurrentUser() currentUser: CurrentUserInfo
+  ) {
+    const cv = await this.cvReadService.findOne(input.tpJobseekerCvId)
+    if (cv.props.userId !== currentUser.userId) {
+      throw new UnauthorizedException(
+        'You are not authorized to create a TpJobseekerCvLanguageRecord for this CV'
+      )
+    }
+
     const props = new TpJobseekerCvEducationRecordEntityProps()
     Object.assign(props, input)
 
@@ -55,8 +73,18 @@ export class TpJobseekerCvEducationRecordsService {
     return await this.api.create(recordToPersist)
   }
 
-  async patch(input: TpJobseekerCvEducationRecordPatchInput) {
+  async patch(
+    input: TpJobseekerCvEducationRecordPatchInput,
+    currentUser: CurrentUserInfo
+  ) {
     const existingEntity = await this.findOne(input.id)
+
+    if (existingEntity.props.userId !== currentUser.userId) {
+      throw new UnauthorizedException(
+        'You are not authorized to update this TpJobseekerCvEducationRecord'
+      )
+    }
+
     const props = existingEntity.props
     const updatesSanitized = deleteUndefinedProperties(input)
     Object.entries(updatesSanitized).forEach(([key, value]) => {
@@ -66,8 +94,18 @@ export class TpJobseekerCvEducationRecordsService {
     await this.api.update(this.mapper.toPersistence(entityToPersist))
   }
 
-  async delete(input: TpJobseekerCvEducationRecordDeleteInput) {
+  async delete(
+    input: TpJobseekerCvEducationRecordDeleteInput,
+    currentUser: CurrentUserInfo
+  ) {
     const existingEntity = await this.findOne(input.id)
+
+    if (existingEntity.props.userId !== currentUser.userId) {
+      throw new UnauthorizedException(
+        'You are not authorized to update this TpJobseekerCvEducationRecord'
+      )
+    }
+
     const recordToDelete = this.mapper.toPersistence(existingEntity)
     await this.api.delete(recordToDelete)
   }
